requirements.txt
obspy>=1.2.2
pandas>=1.4.2
pygmt>=0.5.0
pyyaml>=5.4
